move ResourceCheck into a new file
