+++ huaweicloud/services/acceptance/acceptance.go
@@ -1,21 +1,14 @@
 package acceptance
 
 import (
-	"encoding/json"
 	"fmt"
-	"log"
 	"os"
-	"regexp"
-	"strings"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud"
-	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/config"
 )
 
 var (
@@ -98,222 +91,6 @@ func init() {
 		"huaweicloud": func() (*schema.Provider, error) {
 			return TestAccProvider, nil
 		},
-	}
-}
-
-// ServiceFunc the HuaweiCloud resource query functions.
-type ServiceFunc func(*config.Config, *terraform.ResourceState) (interface{}, error)
-
-// ResourceCheck resource check object
-type ResourceCheck struct {
-	resourceName    string
-	resourceObject  interface{}
-	getResourceFunc ServiceFunc
-	resourceType    string
-}
-
-const (
-	resourceTypeCode   = "resource"
-	dataSourceTypeCode = "dataSource"
-
-	checkAttrRegexpStr = `^\$\{([^\}]+)\}$`
-)
-
-/*
-InitDataSourceCheck build a 'ResourceCheck' object. Only used to check datasource attributes.
-  Parameters:
-    resourceName: The resource name is used to check in the terraform.State.e.g. : huaweicloud_waf_domain.domain_1.
-  Return:
-    *ResourceCheck: ResourceCheck object
-*/
-func InitDataSourceCheck(sourceName string) *ResourceCheck {
-	return &ResourceCheck{
-		resourceName: sourceName,
-		resourceType: dataSourceTypeCode,
-	}
-}
-
-/*
-InitResourceCheck build a 'ResourceCheck' object. The common test methods are provided in 'ResourceCheck'.
-  Parameters:
-    resourceName:    The resource name is used to check in the terraform.State.e.g. : huaweicloud_waf_domain.domain_1.
-    resourceObject:  Resource object, used to check whether the resource exists in HuaweiCloud.
-    getResourceFunc: The function used to get the resource object.
-  Return:
-    *ResourceCheck: ResourceCheck object
-*/
-func InitResourceCheck(resourceName string, resourceObject interface{}, getResourceFunc ServiceFunc) *ResourceCheck {
-	return &ResourceCheck{
-		resourceName:    resourceName,
-		resourceObject:  resourceObject,
-		getResourceFunc: getResourceFunc,
-		resourceType:    resourceTypeCode,
-	}
-}
-
-func parseVariableToName(varStr string) (string, string, error) {
-	var resName, keyName string
-
-	// Check the format of the variable
-	reg, err := regexp.Compile(checkAttrRegexpStr)
-	if err != nil {
-		return resName, keyName, err
-	}
-	mArr := reg.FindStringSubmatch(varStr)
-	if len(mArr) != 2 {
-		return resName, keyName, fmt.Errorf("the type of 'variable' is error, "+
-			"expected ${resource-type.name.field} got %s", varStr)
-	}
-
-	// Get resName and keyName from variable.
-	strs := strings.Split(mArr[1], ".")
-	keyIndex := 2
-	if strs[0] == "data" {
-		keyIndex = 3
-	}
-
-	if len(strs) <= keyIndex {
-		return resName, keyName, fmt.Errorf("attribute field is missing: "+
-			"expected ${resource-type.name.field} got %s", varStr)
-	}
-
-	resName = strings.Join(strs[0:keyIndex], ".")
-	keyName = strings.Join(strs[keyIndex:], ".")
-
-	return resName, keyName, nil
-}
-
-/*
-TestCheckResourceAttrWithVariable validates the variable in state for the given name/key combination.
-  Parameters:
-    resourceName: The resource name is used to check in the terraform.State.
-    key:          The field name of the resource.
-    variable:     The variable name of the value to be checked.
-
-    variable such like ${huaweicloud_waf_certificate.certificate_1.id}
-    or ${data.huaweicloud_waf_policies.policies_2.policies.0.id}
-*/
-func TestCheckResourceAttrWithVariable(resourceName, key, varStr string) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		resName, keyName, err := parseVariableToName(varStr)
-		if err != nil {
-			return err
-		}
-
-		if strings.EqualFold(resourceName, resName) {
-			return fmt.Errorf("meaningless verification: " +
-				"The referenced resource cannot be the current resource")
-		}
-
-		// Get the value based on resName and keyName from the state.
-		rs, ok := s.RootModule().Resources[resName]
-		if !ok {
-			return fmt.Errorf("can't find %s in state: %v", resName, ok)
-		}
-		value := rs.Primary.Attributes[keyName]
-
-		return resource.TestCheckResourceAttr(resourceName, key, value)(s)
-	}
-}
-
-// CheckResourceDestroy check whether resources destroyed in HuaweiCloud.
-func (rc *ResourceCheck) CheckResourceDestroy() resource.TestCheckFunc {
-	if strings.Compare(rc.resourceType, dataSourceTypeCode) == 0 {
-		log.Printf("[WARN] builting a ResourceCheck with 'InitDataSourceCheck', " +
-			"it cannot run CheckResourceDestroy()")
-		return nil
-	}
-
-	return func(s *terraform.State) error {
-		strs := strings.Split(rc.resourceName, ".")
-		resourceType := strs[0]
-
-		if resourceType == "" || resourceType == "data" {
-			return fmt.Errorf("the format of the resource name is invalid, please check your configuration")
-		}
-
-		if rc.getResourceFunc == nil {
-			return fmt.Errorf("the 'getResourceFunc' is nil, please set it during initialization")
-		}
-
-		conf := TestAccProvider.Meta().(*config.Config)
-		for _, rs := range s.RootModule().Resources {
-			if rs.Type != resourceType {
-				continue
-			}
-
-			if _, err := rc.getResourceFunc(conf, rs); err == nil {
-				return fmt.Errorf("failed to destroy the %s resource: %s still exists",
-					resourceType, rs.Primary.ID)
-			}
-		}
-		return nil
-	}
-}
-
-func (rc *ResourceCheck) checkResourceExists(s *terraform.State) error {
-	rs, ok := s.RootModule().Resources[rc.resourceName]
-	if !ok {
-		return fmt.Errorf("can not found the resource or data source in state: %s", rc.resourceName)
-	}
-
-	if rs.Primary.ID == "" {
-		return fmt.Errorf("No id set for the resource or data source: %s", rc.resourceName)
-	}
-	if strings.EqualFold(rc.resourceType, dataSourceTypeCode) {
-		return nil
-	}
-
-	if rc.getResourceFunc == nil {
-		return fmt.Errorf("the 'getResourceFunc' is nil, please set it during initialization")
-	}
-
-	conf := TestAccProvider.Meta().(*config.Config)
-	r, err := rc.getResourceFunc(conf, rs)
-	if err != nil {
-		return fmt.Errorf("checking resource %s %s exists error: %s ",
-			rc.resourceName, rs.Primary.ID, err)
-	}
-
-	if rc.resourceObject != nil {
-		b, err := json.Marshal(r)
-		if err != nil {
-			return fmt.Errorf("marshaling resource %s %s error: %s ",
-				rc.resourceName, rs.Primary.ID, err)
-		}
-		json.Unmarshal(b, rc.resourceObject)
-	} else {
-		log.Printf("[WARN] The 'resourceObject' is nil, please set it during initialization")
-	}
-
-	return nil
-}
-
-// CheckResourceExists check whether resources exist in HuaweiCloud.
-func (rc *ResourceCheck) CheckResourceExists() resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		return rc.checkResourceExists(s)
-	}
-}
-
-/*
-CheckMultiResourcesExists checks whether multiple resources created by count in HuaweiCloud are both existed.
-  Parameters:
-    expCount: the expected number of resources that will be created.
-*/
-func (rc *ResourceCheck) CheckMultiResourcesExists(expCount int) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		var err error
-		for i := 0; i < expCount; i++ {
-			rcCopy := *rc
-			rcCopy.resourceName = fmt.Sprintf("%s.%d", rcCopy.resourceName, i)
-			err = rcCopy.checkResourceExists(s)
-			if err != nil {
-				return err
-			}
-		}
-
-		return nil
 	}
 }
